# contracts/openapi/v1.yaml
...
openapi: 3.1.0
info:
  title: anpfuel community platform
  version: 1.0.0-draft
  description: >
    Base schemas, error envelope, health skeleton and auth profile reference.
    Domain endpoints (directory, official, identity, community) land in later
    tasks, each with contract-first paths added here before code. Writes
    authenticate by anonymous proof of possession (P-256, RFC 9421 profile);
    the exact covered components, encoding and challenge binding freeze in
    P03-T01 (identity profile) and no securityScheme is published before then.
  contact:
    name: anpfuel maintainers
  license:
    name: MIT
    url: https://opensource.org/license/mit
...
paths:
...
  /v1/feedback/ratings:
    post:
      summary: Rate a station/fuel with 1-5 personal stars.
      operationId: rateStation
      tags: [feedback]
      security: []
      description: >
        Session-authenticated write (P22-T02, B-BR-F02): one current
        rating per account/target; equal stars converge idempotently,
        changed stars bump the revision. Returns the rating plus the
        exact count/sum aggregate. Community stars stay separate from
        price confidence (B-BR-C04).
      requestBody:
        required: true
        description: Live session plus target and integer stars.
        content:
          application/json:
            schema:
              type: object
              required: [family_id, access_token, station_id, product, stars]
              properties:
                family_id:
                  type: string
                  format: uuid
                  example: aac027be-d331-40e4-8d63-52ec3b8d2f41
                access_token:
                  type: string
                  example: tok-access
                station_id:
                  type: string
                  format: uuid
                  example: d6c74c23-63db-4c24-a2e5-408cb23bad26
                product:
                  type: string
                  example: GASOLINE_ADDITIVED
                stars:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
      responses:
        '201':
          description: Rating stored with the exact aggregate.
          headers:
            Cache-Control:
              schema: {type: string, example: 'no-store'}
          content:
            application/json:
              schema:
                type: object
                required: [rating, stats]
                properties:
                  rating:
                    $ref: '#/components/schemas/FeedbackRatingReceipt'
                  stats:
                    $ref: '#/components/schemas/FeedbackRatingStats'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/feedback/ratings/remove:
    post:
      summary: Remove my station/fuel rating.
      operationId: removeRating
      tags: [feedback]
      security: []
      description: >
        Tombstones the caller's current rating; unknown keys refuse
        with rating-not-found without touching aggregates.
      requestBody:
        required: true
        description: Live session plus target.
        content:
          application/json:
            schema:
              type: object
              required: [family_id, access_token, station_id, product]
              properties:
                family_id:
                  type: string
                  format: uuid
                  example: aac027be-d331-40e4-8d63-52ec3b8d2f41
                access_token:
                  type: string
                  example: tok-access
                station_id:
                  type: string
                  format: uuid
                  example: d6c74c23-63db-4c24-a2e5-408cb23bad26
                product:
                  type: string
                  example: GASOLINE_ADDITIVED
      responses:
        '200':
          description: Rating removed.
          headers:
            Cache-Control:
              schema: {type: string, example: 'no-store'}
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/feedback/ratings/stats:
    get:
      summary: Read a station/fuel rating aggregate.
      operationId: readRatingStats
      tags: [feedback]
      security: []
      description: >
        Public anonymous read: exact count/sum plus truncated integer
        mean; null mean means no ratings, never an invented score.
      parameters:
        - name: station_id
          in: query
          required: true
          description: Directory station identifier of the rated target.
          schema:
            type: string
            format: uuid
        - name: product
          in: query
          required: true
          description: Catalog fuel product wire name of the rated target.
          schema:
            type: string
      responses:
        '200':
          description: Exact aggregate, zeroed when no rating exists.
          headers:
            Cache-Control:
              schema: {type: string, example: 'no-store'}
          content:
            application/json:
              schema:
                type: object
                required: [stats]
                properties:
                  stats:
                    $ref: '#/components/schemas/FeedbackRatingStats'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/feedback/comments:
    post:
      summary: Submit a station/fuel comment.
      operationId: submitComment
      tags: [feedback]
      security: []
      description: >
        Session-authenticated write (P14-T03, B-BR-F01/F03): the author
        derives server-side from the live session; only active accounts
        write. Tombstoned rows and foreign keys share the 404 shape.
      requestBody:
        required: true
        description: Live session plus target and plain text.
        content:
          application/json:
            schema:
              type: object
              required: [family_id, access_token, station_id, product, text]
              properties:
                family_id:
                  type: string
                  format: uuid
                  example: aac027be-d331-40e4-8d63-52ec3b8d2f41
                access_token:
                  type: string
                  example: tok-access
                station_id:
                  type: string
                  format: uuid
                  example: d6c74c23-63db-4c24-a2e5-408cb23bad26
                product:
                  type: string
                  example: GASOLINE_ADDITIVED
                text:
                  type: string
                  example: Preço bom ⛽
      responses:
        '201':
          description: Comment stored.
          headers:
            Cache-Control:
              schema: {type: string, example: 'no-store'}
          content:
            application/json:
              schema:
                type: object
                required: [comment]
                properties:
                  comment:
                    $ref: '#/components/schemas/FeedbackCommentView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      summary: List station/fuel comments.
      operationId: listComments
      tags: [feedback]
      security: []
      description: >
        Public anonymous read, oldest-first keyset pages. Tombstoned
        rows never appear; author identity stays an opaque alias.
      parameters:
        - name: station_id
          in: query
          required: true
          description: Directory station reviewed.
          schema:
            type: string
            format: uuid
            example: d6c74c23-63db-4c24-a2e5-408cb23bad26
        - name: product
          in: query
          required: true
          description: Catalog fuel product wire name.
          schema:
            type: string
            example: GASOLINE_ADDITIVED
        - name: cursor
          in: query
          description: Opaque page cursor from the previous response.
          schema:
            type: string
        - name: limit
          in: query
          description: Page size, 1 to 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: One bounded page with an opaque cursor.
          headers:
            Cache-Control:
              schema: {type: string, example: 'no-store'}
          content:
            application/json:
              schema:
                type: object
                required: [comments, next_cursor]
                properties:
                  comments:
                    type: array
                    example: []
                    items:
                      $ref: '#/components/schemas/FeedbackCommentView'
                  next_cursor:
                    type: [string, 'null']
                    example: null
              example:
                comments: []
                next_cursor: null
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/feedback/comments/{id}/replies:
    post:
      summary: Reply to a station/fuel comment.
      operationId: replyComment
      tags: [feedback]
      security: []
      description: >
        One-level replies only (F04): the parent fixes the target
        server-side, so substitution across stations, fuels or depths
        refuses. Tombstoned and missing parents share the 404 shape.
      parameters:
        - name: id
          in: path
          required: true
          description: Comment identifier from a prior response.
          schema:
            type: string
            format: uuid
            example: bbd138ce-e442-4f57-a874-3c05d2e16f52
      requestBody:
        required: true
        description: Live session plus the reply text.
        content:
          application/json:
            schema:
              type: object
              required: [family_id, access_token, text]
              properties:
                family_id:
                  type: string
                  format: uuid
                  example: aac027be-d331-40e4-8d63-52ec3b8d2f41
                access_token:
                  type: string
                  example: tok-access
                text:
                  type: string
                  example: Concordo
      responses:
        '201':
          description: Reply stored.
          headers:
            Cache-Control:
              schema: {type: string, example: 'no-store'}
          content:
            application/json:
              schema:
                type: object
                required: [comment]
                properties:
                  comment:
                    $ref: '#/components/schemas/FeedbackCommentView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      summary: List replies of a comment.
      operationId: listReplies
      tags: [feedback]
      security: []
      description: Public anonymous read, oldest-first keyset pages.
      parameters:
        - name: id
          in: path
          required: true
          description: Comment identifier from a prior response.
          schema:
            type: string
            format: uuid
            example: bbd138ce-e442-4f57-a874-3c05d2e16f52
        - name: cursor
          in: query
          description: Opaque page cursor from the previous response.
          schema:
            type: string
        - name: limit
          in: query
          description: Page size, 1 to 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
      responses:
        '200':
          description: One bounded page with an opaque cursor.
          headers:
            Cache-Control:
              schema: {type: string, example: 'no-store'}
          content:
            application/json:
              schema:
                type: object
                required: [comments, next_cursor]
                properties:
                  comments:
                    type: array
                    example: []
                    items:
                      $ref: '#/components/schemas/FeedbackCommentView'
                  next_cursor:
                    type: [string, 'null']
                    example: null
              example:
                comments: []
                next_cursor: null
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /v1/feedback/comments/{id}/votes/remove:
    post:
      summary: Remove my validity vote from a comment.
      operationId: removeVote
      tags: [feedback]
      security: []
      description: >
        Tombstones the caller's live vote on the current revision;
        absent votes converge as a no-op success.
      parameters:
        - name: id
          in: path
          required: true
          description: Comment identifier from a prior response.
          schema:
            type: string
            format: uuid
            example: bbd138ce-e442-4f57-a874-3c05d2e16f52
      requestBody:
        required: true
        description: Live session owning the vote.
        content:
          application/json:
            schema:
              type: object
              required: [family_id, access_token]
              properties:
                family_id:
                  type: string
                  format: uuid
                  example: aac027be-d331-40e4-8d63-52ec3b8d2f41
                access_token:
                  type: string
                  example: tok-access
      responses:
        '200':
          description: Vote removed with the exact tally.
          headers:
            Cache-Control:
              schema: {type: string, example: 'no-store'}
          content:
            application/json:
              schema:
                type: object
                required: [tally]
                properties:
                  tally:
                    $ref: '#/components/schemas/FeedbackTally'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    FeedbackRatingReceipt:
      description: >
        Server-confirmed rating write (P22-T02, B-BR-F02): stored stars
        plus revision and whether this write created the current row.
      type: object
      required: [station_id, product, stars, revision, created]
      properties:
        station_id:
          type: string
          format: uuid
          example: d6c74c23-63db-4c24-a2e5-408cb23bad26
        product:
          type: string
          example: GASOLINE_ADDITIVED
        stars:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        revision:
          type: integer
          minimum: 1
          example: 1
        created:
          type: boolean
          example: true
    FeedbackRatingStats:
      description: >
        Rebuildable rating aggregate (P22-T02, B-BR-F02): exact count
        and star sum per station/product; truncated integer mean in
        thousandths, null when no rating exists. Community stars stay
        separate from price confidence (B-BR-C04).
      type: object
      required: [station_id, product, count, sum]
      properties:
        station_id:
          type: string
          format: uuid
          example: d6c74c23-63db-4c24-a2e5-408cb23bad26
        product:
          type: string
          example: GASOLINE_ADDITIVED
        count:
          type: integer
          minimum: 0
          example: 5
        sum:
          type: integer
          minimum: 0
          example: 21
...